Run all tests under karma.
