+++ test/karma.conf.js
@@ -23,7 +23,7 @@ module.exports = function(config) {
 
     // list of files / patterns to load in the browser
     files: [
-      'test/functional.test.js'
+      'test/*.test.js'
     ],
 
 
@@ -36,7 +36,7 @@ module.exports = function(config) {
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
     preprocessors: {
-      'test/functional.test.js': ['webpack']
+      'test/*.test.js': ['webpack']
     },
